refactor(transformations-editor): move transformation editor components to lib with configuration, testing, and run history functionalities
