+++ apps/app/src/routes/(config)/recordings/row-actions/ViewTransformationRunsDialog.svelte
@@ -5,7 +5,7 @@
 	import { rpc } from '$lib/query';
 	import { createQuery } from '@tanstack/svelte-query';
 	import { HistoryIcon } from 'lucide-svelte';
-	import { Runs } from '../../transformations/editor';
+	import { Runs } from '$lib/components/transformations-editor';
 
 	let { recordingId }: { recordingId: string } = $props();
