+++ apps/app/src/routes/(config)/transformations/CreateTransformationButton.svelte
@@ -8,7 +8,7 @@
 	import { toast } from '$lib/toast';
 	import { createMutation } from '@tanstack/svelte-query';
 	import { PlusIcon } from 'lucide-svelte';
-	import { Editor } from './editor';
+	import { Editor } from '$lib/components/transformations-editor';
 
 	const createTransformation = createMutation(
 		rpc.transformations.mutations.createTransformation.options,
